// src/enums/ERiskLabel.enum.ts
export enum EICQRiskLabel {
  LOW_RISK = 'Risco baixo',
  MODERATE_RISK = 'Risco moderado',
  HIGH_RISK = 'Risco alto',
  VERY_HIGH_RISK = 'Risco muito alto',
}

export enum EIACRiskLabel {
  VERY_LOW = 'Excepcionalmente baixa',
  LOW = 'Baixa',
  NORMAL = 'Ideal',
  MODERATE = 'Moderada',
  HIGH = 'Excesso de gordura',
}

export enum EIMCRiskLabel {
  VERY_LOW = 'Muito abaixo do peso',
  LOW = 'Abaixo do peso',
  NORMAL = 'Normal',
  HIGH = 'Acima do peso',
  OBESITY_1 = 'Obesidade 1',
  OBESITY_2 = 'Obesidade 2 (severa)',
  OBESITY_3 = 'Obesidade 3 (mórbida)',
}